[WIP] read the jpl header in progress
